+++ Shiki/kern_resources.hpp
@@ -28,4 +28,5 @@ enum : uint32_t {
 	SectionOFFLINE = 4,
 	SectionBGRA = 5,
 	SectionNVDA = 6,
+	SectionKILLGVA = 7,
 };